requirements.txt
Django==5.1.6
djangorestframework==3.16.0

celery==5.5.3
redis==6.2.0

nltk==3.9.2
textblob==0.19.0
python-dateutil==2.9.0.post0
drf-yasg==1.21.8